Enable multi-host builds using Docker Swarm
The wrl-ci.yaml file supports both Docker Compose and Docker Swarm
using the docker stack command.

The biggest change was due to a bug in Docker which does not support
--volumes-from when creating a docker container inside a Jenkins
pipeline. If the volume is explicitly given, the docker command
works. To workaround, all docker run commands in the Jenkinsfile
needed to be run using sh instead of using the Jenkins Docker Pipeline
constructs.

The rsync container now has its own rsync network which the postbuild
containers can attach to. This enables the postbuild container to
rsync files to the server without exposing the rsync port.

Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>