Enable runtime testing of build images
This is currently WR specific with links to internal git repos and
LAVA servers.

If the build completes successfully and testing is enabled, the images
are rsynced to the LAVA server and runtime tests like the LAVA smoke
test and oeqa tests are run. The LAVA server is configured to use
internal simics and hardware targets.

Signed-off-by: Wang, Yang (Young) <yang.wang@windriver.com>
Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>